Converted ClientResponse to be an interface Implemented a new type of ClientResponse which should be simpler since it just gets a JavaType passed in, instead of working out how to convert, its just told - convert to this
